+++ src/react/calculator/src/Overview/Overview.css
@@ -1,4 +1,4 @@
-.App2 {
+.Overview {
     width: 100%;
     height: 100vh;
     display: flex;
