+++ src/react/calculator/src/ShoppingList/ShoppingList.css
@@ -1,4 +1,4 @@
-.App2 {
+.ShoppingList {
     width: 100%;
     height: 100vh;
     display: flex;
