+++ src/react/calculator/src/ShoppingList/ShoppingList.js
@@ -1,4 +1,4 @@
-import './App2.css';
+import './ShoppingList.css';
 import useFetch from "react-fetch-hook";
 import { Link } from 'react-router-dom';
 function divider(sharedWith) {
@@ -8,8 +8,8 @@ function divider(sharedWith) {
   else return sharedWith.length + 1
 }
 
-function deleteEntry(billID) {
-  fetch("https://8080-nklsdhbw-webprogramming-ltpyo05qis6.ws-eu81.gitpod.io/api/bills/" + billID, {
+function deleteEntry(shoppingListID) {
+  fetch("https://8080-nklsdhbw-webprogramming-ltpyo05qis6.ws-eu81.gitpod.io/api/shoppingList/" + shoppingListID, {
 
     headers: {
       'Accept': 'application/json',
@@ -26,8 +26,8 @@ function deleteEntry(billID) {
 
 
 
-function App2() {
-  const { isLoading, data } = useFetch("https://8080-nklsdhbw-webprogramming-ltpyo05qis6.ws-eu81.gitpod.io/api/bills");
+function ShoppingList() {
+  const { isLoading, data } = useFetch("https://8080-nklsdhbw-webprogramming-ltpyo05qis6.ws-eu81.gitpod.io/api/shoppingList");
   if (isLoading === false) {
     console.log(Object.keys(data))
     console.log()
@@ -36,21 +36,14 @@ function App2() {
       <div className="App">
         <table>
           <tr>
-            <th>Gläubiger</th>
-            <th>Rechnung</th>
-            <th>Schuldner</th>
-            <th>Datum</th>
-            <th>Betrag</th>
+            <th>Gegenstand</th>
             <th></th>
           </tr>
           {data.map(item => (
             <tr>
-              <td>{item.contributor}</td>
-              <td>{item.billID}</td>
-              <td>{item.sharedWith}</td>
-              <td>{item.datum}</td>
-              <td>{item.amount}</td>
-              <td><button onClick={() => deleteEntry(item.billID)}>Löschen!</button></td>
+              <td>{item.item}</td>
+
+              <td><button onClick={() => deleteEntry(item.shoppingListID)}>Löschen!</button></td>
             </tr>
           ))}
         </table>
@@ -60,4 +53,4 @@ function App2() {
   }
 }
 
-export default App2;
+export default ShoppingList;
