+++ src/react/calculator/src/Splitter/Splitter.css
